+++ mayan/apps/documents/icons.py
@@ -69,10 +69,6 @@ icon_favorite_document_remove = Icon(
 
 # Document files
 
-icon_document_file_cache_purge = Icon(
-    driver_name='fontawesome-dual', primary_symbol='warehouse',
-    secondary_symbol='check'
-)
 icon_document_file_delete = Icon(
     driver_name='fontawesome', symbol='times'
 )
@@ -197,10 +193,6 @@ icon_document_type_setup = icon_document_type
 
 icon_document_version_active = Icon(
     driver_name='fontawesome', symbol='check'
-)
-icon_document_version_cache_purge = Icon(
-    driver_name='fontawesome-dual', primary_symbol='warehouse',
-    secondary_symbol='check'
 )
 icon_document_version_create = Icon(
     driver_name='fontawesome-dual', primary_symbol='code-branch',