+++ mayan/apps/documents/links/document_file_links.py
@@ -6,12 +6,11 @@ from mayan.apps.converter.permissions import (
 from mayan.apps.navigation.classes import Link
 
 from ..icons import (
-    icon_document_file_cache_purge, icon_document_file_delete,
-    icon_document_file_download_quick, icon_document_file_edit,
-    icon_document_file_list, icon_document_file_print,
-    icon_document_file_properties, icon_document_file_return_to_document,
-    icon_document_file_return_list, icon_document_file_preview,
-    icon_document_file_transformations_clear,
+    icon_document_file_delete, icon_document_file_download_quick,
+    icon_document_file_edit, icon_document_file_list,
+    icon_document_file_print, icon_document_file_properties,
+    icon_document_file_return_to_document, icon_document_file_return_list,
+    icon_document_file_preview, icon_document_file_transformations_clear,
     icon_document_file_transformations_clone
 )
 from ..permissions import (
@@ -20,12 +19,6 @@ from ..permissions import (
     permission_document_file_view, permission_document_view
 )
 
-link_document_file_cache_purge = Link(
-    args='resolved_object.pk',
-    icon=icon_document_file_cache_purge,
-    #permissions=(permission_document_file_delete,), tags='dangerous',
-    text=_('Purge cache'), view='documents:document_file_cache_purge'
-)
 link_document_file_delete = Link(
     args='object.pk',
     icon=icon_document_file_delete,