+++ mayan/apps/documents/models/document_file_models.py
@@ -201,7 +201,7 @@ class DocumentFile(
     @method_event(
         event_manager_class=EventManagerMethodAfter,
         event=event_document_file_deleted,
-        action_object='document',
+        target='document',
     )
     def delete(self, *args, **kwargs):
         for page in self.pages.all():
@@ -249,6 +249,13 @@ class DocumentFile(
         first_page = self.pages.first()
         if first_page:
             return first_page.get_api_image_url(*args, **kwargs)
+
+    def get_cache_partitions(self):
+        result = [self.cache_partition]
+        for page in self.file_pages.all():
+            result.append(page.cache_partition)
+
+        return result
 
     def get_intermediate_file(self):
         cache_filename = 'intermediate_file'