+++ mayan/apps/file_caching/icons.py
@@ -3,6 +3,10 @@ from mayan.apps.appearance.classes import Icon
 icon_file_caching = Icon(
     driver_name='fontawesome', symbol='warehouse'
 )
+icon_cache_partition_purge = Icon(
+    driver_name='fontawesome-dual', primary_symbol='warehouse',
+    secondary_symbol='check'
+)
 icon_cache_purge = Icon(
     driver_name='fontawesome-dual', primary_symbol='warehouse',
     secondary_symbol='check'