+++ mayan/apps/file_caching/links.py
@@ -1,9 +1,12 @@
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.navigation.classes import Link
+from mayan.apps.navigation.utils import get_content_type_kwargs_factory
 from mayan.apps.storage.classes import DefinedStorage
 
-from .icons import icon_cache_purge, icon_file_caching
+from .icons import (
+    icon_cache_partition_purge, icon_cache_purge, icon_file_caching
+)
 from .permissions import permission_cache_purge, permission_cache_view
 
 
@@ -19,6 +22,12 @@ def condition_valid_storage(context):
 link_caches_list = Link(
     icon=icon_file_caching, permissions=(permission_cache_view,),
     text=_('File caches'), view='file_caching:cache_list'
+)
+link_cache_partition_purge = Link(
+    icon=icon_cache_partition_purge, kwargs=get_content_type_kwargs_factory(
+        variable_name='resolved_object'
+    ), permissions=(permission_cache_purge,), text=_('Purge cache'),
+    view='file_caching:cache_partitions_purge'
 )
 link_cache_purge = Link(
     condition=condition_valid_storage, icon=icon_cache_purge,