+++ mayan/apps/file_caching/permissions.py
@@ -4,6 +4,9 @@ from mayan.apps.permissions import PermissionNamespace
 
 namespace = PermissionNamespace(label=_('File caching'), name='file_caching')
 
+permission_cache_partition_purge = namespace.add_permission(
+    label=_('Purge an object cache'), name='file_caching_cache_partition_purge'
+)
 permission_cache_purge = namespace.add_permission(
     label=_('Purge a file cache'), name='file_caching_cache_purge'
 )